Added switch to suppress validation via batch account and cmd-line help
